Move more code generation into corresponding module
